app/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core == 1.15.0
azure-functions == 1.7.2
azure-identity == 1.6.0
azure-keyvault-secrets == 4.3.0
azure-mgmt-resource == 18.0.0
azure-mgmt-containerinstance == 7.0.0
python-decouple == 3.4
PyGithub == 1.55
